requirements.txt
filelock==3.13.3
fsspec==2024.3.1
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
pandas==2.2.1
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
sympy==1.12
threadpoolctl==3.4.0
torch==2.2.2
torchvision==0.17.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
